# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Memory-based General Operating System API with comprehensive swagger documentation",
        "title": "MemGOS API",
        "contact": {
            "name": "MemGOS API Support",
            "url": "https://github.com/pnocera/MemGOS",
            "email": "support@memgos.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Chat with AI using memory-enhanced responses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat with AI",
                "parameters": [
                    {
                        "description": "Chat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/memories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all memories from a memory cube",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Get All Memories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Memory cube ID",
                        "name": "mem_cube_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MemoryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add new memories to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Add Memory",
                "parameters": [
                    {
                        "description": "Memory creation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.MemoryCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SimpleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/search": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search through memories using semantic search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memories"
                ],
                "summary": "Search Memories",
                "parameters": [
                    {
                        "description": "Search parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.ChatRequest": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "context": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "max_tokens": {
                    "type": "integer",
                    "example": 1000
                },
                "query": {
                    "type": "string",
                    "example": "What is the latest update?"
                },
                "top_k": {
                    "type": "integer",
                    "example": 5
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "api.ChatResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.MemoryCreate": {
            "type": "object",
            "properties": {
                "doc_path": {
                    "type": "string",
                    "example": "/path/to/document.txt"
                },
                "mem_cube_id": {
                    "type": "string",
                    "example": "cube123"
                },
                "memory_content": {
                    "type": "string",
                    "example": "This is a memory content"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Message"
                    }
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "api.MemoryResponse": {
            "type": "object"
        },
        "api.Message": {
            "type": "object",
            "required": [
                "content",
                "role"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello, how can I help you?"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "api.SearchRequest": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "install_cube_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "cube123",
                        "cube456"
                    ]
                },
                "query": {
                    "type": "string",
                    "example": "How to implement a feature?"
                },
                "top_k": {
                    "type": "integer",
                    "example": 5
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "api.SearchResponse": {
            "type": "object"
        },
        "api.SimpleResponse": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token authentication. Use 'Bearer {token}' format.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}